Render sprite quality tiers from source scale
Refactor the sprite renderer publish path so quality-specific sheets can be rendered at a target-relative source scale before atlas packing. Fractional render_scale values are now supported, target commands accept --quality-scale / --downsample, and AdapterTarget applies quality scale relative to the target's normal render scale. Add focused tests covering fractional source rendering and relative quality scaling.

Prompt summary: implement a sprite/texture quality-generation refactor so low, medium, and potato actor assets are generated before packing instead of resizing already-packed atlas pages; prefer the elegant source-rendered fix even if larger, and allow quality tiers to pack differently.

+++ ambition_sprite2d_renderer/registry/config.py
@@ -36,7 +36,7 @@ class RenderConfig:
     # and takes only ASPECT from the frame, so this is pure anti-pixelation:
     # higher resolution under the same on-screen quad, no gameplay change.
     # Default 2 because most sheets are upscaled in game and read soft at 1x.
-    render_scale: int = 2
+    render_scale: float = 2.0
     background: str = "transparent"
     sheet_background: str = "transparent"
     border: int = 0
